Create Static & Dynamic Binding.cpp
In this example, we have two classes Animal and Dog. Dog inherits from Animal and overrides the speak() function. In main(), we create two objects, animal1 and animal2, of type Animal*, but animal2 is actually a Dog object.

When we call speak() on animal1, it is determined at compile time that animal1 is an Animal object, so the Animal::speak() function is called. This is an example of static binding.

When we call speak() on animal2, it is determined at runtime that animal2 is actually a Dog object, so the Dog::speak() function is called instead. This is an example of dynamic binding, as the function to be called is determined based on the dynamic type of the object pointed to by the animal2 pointer.