Create Virtual Functions.cpp
In this code, we have three classes: Base, Der1, and Der2. Der1 and Der2 are derived classes of Base.

In the main function, two pointers ptr1 and ptr2 are created and assigned addresses of objects of Der1 and Der2 classes respectively. Then, the show() function is called using these pointers.

However, the output will be Base class in both cases. This is because the show() function is not declared as virtual in the Base class. When a function is not declared as virtual, the function call is resolved at compile-time based on the type of the pointer, not the type of the object that it points to.

In other words, even though ptr1 points to an object of Der1, the show() function that is called is the one declared in Base class because ptr1 is of type Base*. Similarly, ptr2 is also of type Base*, so the show() function called is the one in Base class.