Create type_info in RTTI.cpp
The type_info class in C++ provides runtime type information (RTTI) about the class or type of an object. It has several member functions that can be used to obtain information about a type. Here are some of the most commonly used member functions:

name(): Returns a string representing the name of the type.

operator==() and operator!=(): These operators can be used to compare two type_info objects for equality or inequality.

before(): This function is used to compare two type_info objects to determine which one comes first in the implementation-defined order.

hash_code(): Returns a hash code value for the type.

raw_name(): Returns a string representing a mangled name that uniquely identifies the type.