Update RTTI (type_info).cpp
This C++ code defines two classes: Shape and Square. Shape is an abstract class that contains a pure virtual function called "draw" which has no implementation. Square is a subclass of Shape that overrides the draw function to provide its own implementation.

In the main function, an object of the Square class is created and its address is assigned to a Shape pointer. The code then uses the typeid operator to get information about the type of the object that the Shape pointer is pointing to. This is done by dereferencing the Shape pointer and passing the resulting object to the typeid function. The resulting type_info object contains information about the type of the object, including its name and a hash code.

The code then prints the name and hash code of the object's type using the name and hash_code functions of the type_info object. Finally, the code compares the object's type with the type of the Square class using the == operator and prints a message if they match.

Overall, this code demonstrates how to use the typeid operator to determine the type of an object at runtime, which can be useful in situations where the type of an object is not known until runtime or when dealing with polymorphic classes.